+++ ui/src/app/js/controllers/container/container_controller.js
@@ -109,10 +109,6 @@ angular.module("app").controller('ContainerController', function($scope, $routeP
       });
     });
   }; 
-
-  $scope.send = function(){
-    $scope.command = ""; 
-  };
  
   $scope.initConsole = function(){
     if(dataStream){
@@ -121,44 +117,19 @@ angular.module("app").controller('ContainerController', function($scope, $routeP
     dataStream = DockerResource.getConsoleContainer({id:$routeParams.id});
     if(dataStream){
       dataStream.onMessage(function(message){
+        $scope.console += message.data;
         console.log(message.data);
-        message.data.split('').every(function(character){
-          var charCode = character.charCodeAt(0); 
-          if(charCode !== 0){
-            if(charCode == 8){
-              //backspace
-              $scope.console = $scope.console.replace(/.(\x1B\[[^a-z]*[a-z])?$/gi,'');
-
-              // \x1B\[([^a-z]+)[a-z]/gi
-            }else{
-              //add
-              $scope.console += character;
-            }
-          }
-          return charCode !== 0; 
-        });
         $('.console .output .content').stop().animate({scrollTop: $('.console .output .content')[0].scrollHeight},100);
       });
     }     
   };
 
   $scope.handleConsoleKeyboard = function($event){
-    var character;    
-    if($event.ctrlKey && $event.keyCode >= 65 && $event.keyCode <= 95){
-      character = String.fromCharCode($event.keyCode - 64);
-    }else if($event.key.length == 1){
-      character = $event.key;
-    }else if($event.keyCode >= 37 && $event.keyCode <= 40){
-      if($event.keyCode == 38){
-        character = String.fromCharCode($event.keyCode - 22);
-      }
-    }else if($event.key == 'Shift'){
-      // unhandled arrow keys in html
-    }else{
-      character = String.fromCharCode($event.keyCode);
-    }
-    if(dataStream){
+    var character = extractKey($event); 
+
+    if(dataStream && character !== false){
       dataStream.send(character);   
+      $scope.command = ""; 
     }
   };