+++ ui/src/app/js/controllers/container/containers_controller.js
@@ -7,7 +7,6 @@ angular.module("app").controller('ContainersController', function($scope, Docker
 	$scope.reverse = false;
 
 	$scope.order = function(predicate){
-		console.log('order');
 		$scope.reverse = ($scope.predicate === predicate) ? !$scope.reverse : false;
 		$scope.predicate = predicate;
 	};