+++ ui/src/app/js/controllers/image/images_controller.js
@@ -5,7 +5,6 @@ angular.module("app").controller('ImagesController', function($scope, DockerReso
 	$scope.reverse = false;
 
 	$scope.order = function(predicate){
-		console.log('order');
 		$scope.reverse = ($scope.predicate === predicate) ? !$scope.reverse : false;
 		$scope.predicate = predicate;
 	};